move token utils
