+++ test/unit/test_sciencemesh.py
@@ -93,7 +93,7 @@ def test_create_ocm_share_uses_token_claims_for_owner_and_sender(
 
 
 def test_create_ocm_share_falls_back_name_to_email(sciencemesh_vre, mock_token_user):
-    from app.services.token_utils import TokenUser
+    from app.vres.utils.token_utils import TokenUser
 
     mock_token_user.return_value = TokenUser(
         email="jdoe@example.com",
